make basic tui work
