+++ src/tui/common.rs
@@ -1,7 +1,7 @@
 use ratatui::style::Color;
 
 pub const SECTION_TITLE_FG_COLOR: Color = Color::from_u32(0x151515);
-pub const PRIMARY_COLOR: Color = Color::from_u32(0xffaff3);
+pub const PRIMARY_COLOR: Color = Color::from_u32(0xcc8cc2);
 pub const SECONDARY_COLOR: Color = Color::from_u32(0xa6f0fc);
 pub const MESSAGE_COLOR: Color = Color::from_u32(0xd3869b);
 pub const INACTIVE_PANE_TITLE_BG_COLOR: Color = Color::from_u32(0xbdae93);
